backend/learnervoice/requirements.txt
django==1.8.5
mysqlclient==1.3.6
django-haystack==2.4.0
djangorestframework==3.2.4
